requirements.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
gevent==1.2.2
greenlet==0.4.12
idna==2.6
lxml==4.1.1
ordereddict==1.1
PyMySQL==0.7.11
python-wordpress-xmlrpc==2.3
requests==2.18.4
urllib3==1.22
wordpress-api==1.2.4
